requirements.txt
# Core API and server
fastapi
uvicorn[standard]

# Data validation and utilities
pydantic
python-jose[cryptography]
python-multipart

# Database (if you are using one)
SQLAlchemy>=2.0
psycopg2-binary
python-dotenv

# QR code generation
qrcode[pil]
pydantic[email]
PyJWT
fastapi 
uvicorn 
sqlalchemy 
asyncpg 
PyJWT 
python-dotenv
